+++ mithril-common/src/crypto_helper/types/wrappers.rs
@@ -1,7 +1,8 @@
 use kes_summed_ed25519::kes::Sum6KesSig;
 use mithril_stm::{
     AggregateSignature, AggregateVerificationKey, AggregateVerificationKeyForConcatenation,
-    SingleSignature, VerificationKeyProofOfPossessionForConcatenation,
+    AncillaryProverData, AncillaryVerifierData, SingleSignature,
+    VerificationKeyProofOfPossessionForConcatenation,
 };
 #[cfg(feature = "future_snark")]
 use mithril_stm::{AggregateVerificationKeyForSnark, VerificationKeyForSnark};
@@ -52,13 +53,19 @@ pub type ProtocolAggregateVerificationKeyForSnark =
 /// Wrapper of [MKProof] to add serialization utilities.
 pub type ProtocolMkProof = ProtocolKey<MKMapProof<BlockRange>>;
 
+/// Wrapper of [MithrilStm:AncillaryProverData](enum@AncillaryProverData) to add serialization utilities.
+pub type ProtocolAncillaryProverData = ProtocolKey<AncillaryProverData>;
+
+/// Wrapper of [MithrilStm:AncillaryVerifierData](enum@AncillaryVerifierData) to add serialization utilities.
+pub type ProtocolAncillaryVerifierData = ProtocolKey<AncillaryVerifierData>;
+
 impl_codec_and_type_conversions_for_protocol_key!(
     json_hex_codec => AggregateSignature<ProtocolMembershipDigest>, ed25519_dalek::VerifyingKey, ed25519_dalek::SigningKey, AggregateVerificationKeyForConcatenation<ProtocolMembershipDigest>,
         MKProof, VerificationKeyProofOfPossessionForConcatenation, Sum6KesSig, OpCert, SingleSignature
 );
 
 impl_codec_and_type_conversions_for_protocol_key!(
-    bytes_hex_codec => ed25519_dalek::Signature
+    bytes_hex_codec => ed25519_dalek::Signature, AncillaryProverData, AncillaryVerifierData
 );
 
 #[cfg(feature = "future_snark")]