+++ mithril-common/src/test/builder/certificate_chain_builder.rs
@@ -3,7 +3,7 @@ use std::collections::{BTreeSet, HashMap};
 use std::iter::repeat_n;
 use std::ops::{Deref, DerefMut};
 
-use mithril_stm::AggregateSignatureType;
+use mithril_stm::{AggregateSignatureType, AncillaryProofInput};
 
 use crate::{
     certificate_chain::CertificateGenesisProducer,
@@ -17,7 +17,7 @@ use crate::{
     },
     test::{
         builder::{MithrilFixture, MithrilFixtureBuilder, SignerFixture},
-        double::fake_data,
+        double::{Dummy, fake_data},
     },
 };
 
@@ -573,11 +573,12 @@ impl<'a> CertificateChainBuilder<'a> {
             .filter_map(|s| s.protocol_signer.sign(certificate.signed_message.as_bytes()))
             .collect::<Vec<_>>();
         let clerk = CertificateChainBuilder::compute_clerk_for_signers(&fixture.signers_fixture());
-        let multi_signature = clerk
+        let (multi_signature, _ancillary_verifier_data) = clerk
             .aggregate_signatures_with_type(
                 &single_signatures,
                 certificate.signed_message.as_bytes(),
                 self.aggregate_signature_type,
+                AncillaryProofInput::dummy(),
             )
             .unwrap();
         certificate.signature = CertificateSignature::MultiSignature(